feat: ポーズアイコンの実装
