+++ src/features/ui/Icon/PauseIcon.tsx
@@ -1,0 +1,21 @@
+type Props = {
+  size?: number;
+  width?: number;
+  height?: number;
+};
+export function PauseIcon({ size, width, height }: Props) {
+  return (
+    <svg
+      width={size || width || 40}
+      height={size || height || 40}
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M16 19C15.45 19 14.9793 18.8043 14.588 18.413C14.1967 18.0217 14.0007 17.5507 14 17V7C14 6.45 14.196 5.97933 14.588 5.588C14.98 5.19667 15.4507 5.00067 16 5C16.55 5 17.021 5.196 17.413 5.588C17.805 5.98 18.0007 6.45067 18 7V17C18 17.55 17.8043 18.021 17.413 18.413C17.0217 18.805 16.5507 19.0007 16 19ZM8 19C7.45 19 6.97933 18.8043 6.588 18.413C6.19667 18.0217 6.00067 17.5507 6 17V7C6 6.45 6.196 5.97933 6.588 5.588C6.98 5.19667 7.45067 5.00067 8 5C8.55 5 9.021 5.196 9.413 5.588C9.805 5.98 10.0007 6.45067 10 7V17C10 17.55 9.80433 18.021 9.413 18.413C9.02167 18.805 8.55067 19.0007 8 19Z"
+        fill="white"
+      />
+    </svg>
+  );
+}
