+++ src/features/ui/Icon/SettingsIcon.tsx
@@ -1,0 +1,21 @@
+type Props = {
+  size?: number;
+  width?: number;
+  height?: number;
+};
+export function SettingsIcon({ size, width, height }: Props) {
+  return (
+    <svg
+      width={size || width || 40}
+      height={size || height || 40}
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M10.125 22C9.875 22 9.65833 21.9167 9.475 21.75C9.29167 21.5833 9.18333 21.375 9.15 21.125L8.85 18.8C8.63333 18.7167 8.42933 18.6167 8.238 18.5C8.04667 18.3833 7.859 18.2583 7.675 18.125L5.5 19.025C5.26667 19.125 5.03333 19.1377 4.8 19.063C4.56667 18.9883 4.38333 18.8423 4.25 18.625L2.4 15.4C2.26667 15.1833 2.225 14.95 2.275 14.7C2.325 14.45 2.45 14.25 2.65 14.1L4.525 12.675C4.50833 12.5583 4.5 12.446 4.5 12.338V11.663C4.5 11.5543 4.50833 11.4417 4.525 11.325L2.65 9.9C2.45 9.75 2.325 9.55 2.275 9.3C2.225 9.05 2.26667 8.81667 2.4 8.6L4.25 5.375C4.38333 5.15833 4.56667 5.01267 4.8 4.938C5.03333 4.86333 5.26667 4.87567 5.5 4.975L7.675 5.875C7.85833 5.74167 8.05 5.61667 8.25 5.5C8.45 5.38333 8.65 5.28333 8.85 5.2L9.15 2.875C9.18333 2.625 9.29167 2.41667 9.475 2.25C9.65833 2.08333 9.875 2 10.125 2H13.875C14.125 2 14.3417 2.08333 14.525 2.25C14.7083 2.41667 14.8167 2.625 14.85 2.875L15.15 5.2C15.3667 5.28333 15.571 5.38333 15.763 5.5C15.955 5.61667 16.1423 5.74167 16.325 5.875L18.5 4.975C18.7333 4.875 18.9667 4.86233 19.2 4.937C19.4333 5.01167 19.6167 5.15767 19.75 5.375L21.6 8.6C21.7333 8.81667 21.775 9.05 21.725 9.3C21.675 9.55 21.55 9.75 21.35 9.9L19.475 11.325C19.4917 11.4417 19.5 11.5543 19.5 11.663V12.337C19.5 12.4457 19.4833 12.5583 19.45 12.675L21.325 14.1C21.525 14.25 21.65 14.45 21.7 14.7C21.75 14.95 21.7083 15.1833 21.575 15.4L19.725 18.6C19.5917 18.8167 19.4043 18.9667 19.163 19.05C18.9217 19.1333 18.684 19.125 18.45 19.025L16.325 18.125C16.1417 18.2583 15.95 18.3833 15.75 18.5C15.55 18.6167 15.35 18.7167 15.15 18.8L14.85 21.125C14.8167 21.375 14.7083 21.5833 14.525 21.75C14.3417 21.9167 14.125 22 13.875 22H10.125ZM12.05 15.5C13.0167 15.5 13.8417 15.1583 14.525 14.475C15.2083 13.7917 15.55 12.9667 15.55 12C15.55 11.0333 15.2083 10.2083 14.525 9.525C13.8417 8.84167 13.0167 8.5 12.05 8.5C11.0667 8.5 10.2373 8.84167 9.562 9.525C8.88667 10.2083 8.54933 11.0333 8.55 12C8.55 12.9667 8.88767 13.7917 9.563 14.475C10.2383 15.1583 11.0673 15.5 12.05 15.5Z"
+        fill="white"
+      />
+    </svg>
+  );
+}
